requirements.txt
pandas
jupyter
rank-bm25
fastapi
uvicorn
streamlit
google-genai
python-dotenv